issue-039 FeedList(검색,프로필 공유) 수정
