+++ app/src/main/java/com/boostcamp/dreampicker/view/feedList/FeedListAdapter.java
@@ -1,4 +1,4 @@
-package com.boostcamp.dreampicker.view.profile;
+package com.boostcamp.dreampicker.view.feedList;
 
 import android.view.LayoutInflater;
 import android.view.View;
@@ -6,18 +6,19 @@ import android.view.ViewGroup;
 
 import com.boostcamp.dreampicker.R;
 import com.boostcamp.dreampicker.databinding.ItemFeedListBinding;
+import com.boostcamp.dreampicker.model.Feed;
 import com.boostcamp.dreampicker.view.adapter.BaseRecyclerViewAdapter;
 
 import androidx.annotation.NonNull;
 import androidx.databinding.DataBindingUtil;
 import androidx.recyclerview.widget.RecyclerView;
 
-public class FeedListAdapter extends BaseRecyclerViewAdapter<String, FeedListAdapter.ViewHolder> {
+public class FeedListAdapter extends BaseRecyclerViewAdapter<Feed, FeedListAdapter.ViewHolder> {
 
     @Override
     protected void onBindView(ViewHolder holder, int position) {
 
-        holder.binding.ivTemp.setImageResource(R.drawable.skin);
+        holder.binding.setFeed(itemList.get(position));
     }
 
     @NonNull
