+++ app/src/main/res/values/colors.xml
@@ -4,6 +4,7 @@
     <color name="colorPrimaryDark">#47B985</color>
     <color name="colorAccent">#FD8B44</color>
     <color name="colorBlack">#000000</color>
+    <color name="colorBlackTransParent">#33000000</color>
     <color name="colorGray">#D2D2D7</color>
     <color name="colorWhite">#FFFFFF</color>
     <color name="colorWhiteBackground">#FEFEFE</color>
